Fix: users.telegram_id -> BigInteger (supporto ID Telegram > int32)

+++ src/database_async.py
@@ -6,7 +6,7 @@ import logging
 from datetime import datetime
 from typing import Optional, List, Dict, Any
 from sqlalchemy.ext.asyncio import create_async_engine, async_sessionmaker, AsyncSession
-from sqlalchemy import select, text as sql_text, Column, Integer, String, Float, DateTime, Boolean, Text, ForeignKey
+from sqlalchemy import select, text as sql_text, Column, Integer, BigInteger, String, Float, DateTime, Boolean, Text, ForeignKey
 from sqlalchemy.ext.declarative import declarative_base
 
 logger = logging.getLogger(__name__)
@@ -20,7 +20,7 @@ class User(Base):
     __tablename__ = 'users'
     
     id = Column(Integer, primary_key=True)
-    telegram_id = Column(Integer, unique=True, nullable=False, index=True)
+    telegram_id = Column(BigInteger, unique=True, nullable=False, index=True)
     username = Column(String(100))
     first_name = Column(String(100))
     last_name = Column(String(100))
